#TBSFMD-112: Admin can create and disable customers.  #TBSFMD-117: Admin can create and disable organizers.

+++ src/main/java/com/uol/smqa/repository/AdminRepository.java
@@ -2,7 +2,7 @@ package com.uol.smqa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uol.smqa.model.Admin;
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 	
 
 }
@@ -10,3 +10,4 @@ public interface AdminRepository extends JpaRepository<Admin, Long> {
 
 
 
+//changed Long to Integer
